requirements.txt
plotly-express
streamlit-aggrid
streamlit-option-menu
folium
streamlit_folium
pandas 
numpy
matplotlib
seaborn
openpyxl
xlsxwriter
plotly-express
streamlit-aggrid
streamlit-option-menu
pandas 
numpy
matplotlib
seaborn
openpyxl
xlsxwriter
fpdf
python-docx
docx

pyxlsb
